backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor>=3.6.0
pymongo>=4.10.0
bcrypt>=4.2.0
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.5.0
email-validator>=2.0.0
fastapi-mail>=1.4.1
python-multipart>=0.0.6
httpx>=0.24.0
qrcode[pil]>=7.4.2
pillow>=10.0.0
requests>=2.31.0
cloudinary>=1.36.0

